nothing-ever-changes/src/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.3.0
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1
pycryptodome
ecdsa
gmpy2